+++ src/processors/mixed.ts
@@ -1,4 +1,3 @@
-// @ts-expect-error walk is not in d.ts
 import { walk } from 'svelte/compiler';
 import type { Ast, TemplateNode } from 'svelte/types/compiler/interfaces.d';
 import type { PluginOptions } from '../types';
@@ -32,44 +31,55 @@ const updateSelectorBoundaries = (
  * @param processor The CSS Module Processor
  */
 const parser = (processor: Processor): void => {
-  walk(processor.ast, {
-    enter(node: TemplateNode) {
+  const ast = (processor.ast as unknown) as TemplateNode;
+  walk(ast, {
+    enter(baseNode) {
+      const node = baseNode as TemplateNode;
       if (node.type === 'Script' || node.type === 'Fragment') {
         this.skip();
       }
 
       if (node.type === 'Selector') {
-        const classSelectors = node.children.filter((item) => item.type === 'ClassSelector');
+        const classSelectors = node.children
+          ? node.children.filter((item: { type: string }) => item.type === 'ClassSelector')
+          : [];
         if (classSelectors.length > 0) {
           let selectorBoundaries: Array<Boundaries> = [];
           let start = 0;
           let end = 0;
 
-          node.children.forEach((item, index) => {
-            if (!item.start && start > 0) {
-              selectorBoundaries = updateSelectorBoundaries(selectorBoundaries, start, end);
-              start = 0;
-              end = 0;
-            } else {
-              let hasPushed = false;
-              if (end !== item.start) {
-                start = item.start;
-                end = item.end;
-              } else {
-                selectorBoundaries = updateSelectorBoundaries(selectorBoundaries, start, item.end);
-                hasPushed = true;
+          if (node.children) {
+            node.children.forEach((item: { start: number; end: number }, index: number) => {
+              if (!item.start && start > 0) {
+                selectorBoundaries = updateSelectorBoundaries(selectorBoundaries, start, end);
                 start = 0;
                 end = 0;
+              } else {
+                let hasPushed = false;
+                if (end !== item.start) {
+                  start = item.start;
+                  end = item.end;
+                } else {
+                  selectorBoundaries = updateSelectorBoundaries(
+                    selectorBoundaries,
+                    start,
+                    item.end
+                  );
+                  hasPushed = true;
+                  start = 0;
+                  end = 0;
+                }
+                if (hasPushed === false && node.children && index === node.children.length - 1) {
+                  selectorBoundaries = updateSelectorBoundaries(selectorBoundaries, start, end);
+                }
               }
-              if (hasPushed === false && index === node.children.length - 1) {
-                selectorBoundaries = updateSelectorBoundaries(selectorBoundaries, start, end);
-              }
-            }
-          });
+            });
+          }
 
           selectorBoundaries.forEach((boundary) => {
             const hasClassSelector = classSelectors.filter(
-              (item) => boundary.start <= item.start && boundary.end >= item.end
+              (item: { start: number; end: number }) =>
+                boundary.start <= item.start && boundary.end >= item.end
             );
             if (hasClassSelector.length > 0) {
               processor.magicContent.appendLeft(boundary.start, ':global(');