+++ src/processors/scoped.ts
@@ -1,4 +1,3 @@
-// @ts-expect-error walk is not in d.ts
 import { walk } from 'svelte/compiler';
 import type { Ast, TemplateNode } from 'svelte/types/compiler/interfaces.d';
 import type { PluginOptions } from '../types';
@@ -9,8 +8,10 @@ import Processor from './processor';
  * @param processor The CSS Module Processor
  */
 const parser = (processor: Processor): void => {
-  walk(processor.ast, {
-    enter(node: TemplateNode) {
+  const ast = (processor.ast as unknown) as TemplateNode;
+  walk(ast, {
+    enter(baseNode) {
+      const node = baseNode as TemplateNode;
       if (node.type === 'Script' || node.type === 'Fragment') {
         this.skip();
       }